Add .desktop .appdata.xml and .png files
To integrate in linux desktop we've to implement the freedesktop spec,
so we need to provide a .desktop file, an appdata.xml for "stores" and
at least an icon with the same name as the app ID.

I've added this information in the data directory and it will be
installed when generating the release with the distZip task.

These files are also useful for flatpak packaging.